+++ wallet/wallet-ports/wallet-eventlistener-kafka/src/main/kotlin/co/nilin/opex/wallet/ports/kafka/listener/model/AddCurrencyEvent.kt
@@ -1,7 +1,9 @@
 package co.nilin.opex.wallet.ports.kafka.listener.model
 
+import java.math.BigDecimal
+
 data class AddCurrencyEvent(
     val name: String,
     val symbol: String,
-    val precision: Double
+    val precision: BigDecimal
 ) : AdminEvent()
